requirements.txt
Django==2.2.10
django-crispy-forms==1.7.2
Pillow==6.2.2
pytz==2019.3
sqlparse==0.3.0
python-decouple==3.3
dj-database-url==0.5.0